requirements.txt
psutil>=5.9.0
pytest>=7.0.0
numpy>=1.21.0
faiss-cpu==1.7.4